+++ src/java/de/bno/jellysplush/data/Position.java
@@ -2,8 +2,8 @@ package de.bno.jellysplush.data;
 
 public class Position {
 
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 
 	public Position() {
 		this(0, 0);
@@ -15,19 +15,19 @@ public class Position {
 		this.y = y;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 }
